correção do sistema arquivos
caso o name retornado não seja um arquivo ele não executa a funçaõ
unlink e retorna falste.

+++ _instalar/baseSistema/jquerycms/lib/sistema/arquivos.php
@@ -84,7 +84,7 @@ class arquivos {
     }
 
     public static function deletar($filename) {
-        if (self::existe($filename))
+        if (self::existe($filename) && is_file($filename))
             return unlink($filename);
         else
             return false;
